backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
django-cors-headers==3.14.0
Pillow==9.5.0
